// src/server/model/itemTransactionRegisterReportLine.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TransactionInfo } from './transactionInfo';

export interface ItemTransactionRegisterReportLine { 
    date?: Date;
    name?: string;
    voucher?: string;
    productCode?: string;
    unit?: string;
    quantity?: number;
    amount?: number;
    transaction?: TransactionInfo;
}